+++ src/main/kotlin/it/czerwinski/intellij/wavefront/editor/actions/ToggleAxesAction.kt
@@ -19,11 +19,8 @@ package it.czerwinski.intellij.wavefront.editor.actions
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.Toggleable
 import com.intellij.openapi.project.DumbAware
-import it.czerwinski.intellij.wavefront.editor.model.UpVector
 
-abstract class UpVectorAction(
-    private val upVector: UpVector
-) : ObjPreviewFileEditorAction(), DumbAware, Toggleable {
+class ToggleAxesAction : ObjPreviewFileEditorAction(), DumbAware, Toggleable {
 
     override fun update(event: AnActionEvent) {
         val editor = findObjPreviewFileEditor(event)
@@ -31,7 +28,7 @@ abstract class UpVectorAction(
         event.presentation.isEnabled = editor != null
 
         if (editor != null) {
-            Toggleable.setSelected(event.presentation, editor.upVector === upVector)
+            Toggleable.setSelected(event.presentation, editor.isShowingAxes)
         }
     }
 
@@ -39,8 +36,8 @@ abstract class UpVectorAction(
         val editor = findObjPreviewFileEditor(event)
 
         if (editor != null) {
-            editor.triggerUpVectorChange(upVector)
-            Toggleable.setSelected(event.presentation, true)
+            Toggleable.setSelected(event.presentation, !Toggleable.isSelected(event.presentation))
+            editor.toggleAxes()
         }
     }
 }